Move metrics csv marshalling to separate type
