+++ ingestor/transform/csv_test.go
@@ -11,153 +11,9 @@ import (
 	v1 "buf.build/gen/go/opentelemetry/opentelemetry/protocolbuffers/go/opentelemetry/proto/collector/logs/v1"
 	"github.com/Azure/adx-mon/collector/logs/types"
 	"github.com/Azure/adx-mon/pkg/otlp"
-	"github.com/Azure/adx-mon/pkg/prompb"
 	"github.com/stretchr/testify/require"
 	"google.golang.org/protobuf/encoding/protojson"
 )
-
-func TestMarshalCSV(t *testing.T) {
-	ts := &prompb.TimeSeries{
-		Labels: []*prompb.Label{
-			{
-				Name:  []byte("__name__"),
-				Value: []byte("__redis__"),
-			},
-			{
-				Name:  []byte("measurement"),
-				Value: []byte("used_cpu_user_children"),
-			},
-			{
-				Name:  []byte("hostname"),
-				Value: []byte("host_1"),
-			},
-			{
-				Name:  []byte("region"),
-				Value: []byte("eastus"),
-			},
-		},
-
-		Samples: []*prompb.Sample{
-			{
-				Timestamp: 1669112524001,
-				Value:     0,
-			},
-			{
-				Timestamp: 1669112525002,
-				Value:     1,
-			},
-			{
-				Timestamp: 1669112526003,
-				Value:     2,
-			},
-		},
-	}
-
-	var b bytes.Buffer
-	w := NewCSVWriter(&b, nil)
-	err := w.MarshalTS(ts)
-	require.NoError(t, err)
-	require.Equal(t, `2022-11-22T10:22:04.001Z,-9070404444212865161,"{""measurement"":""used_cpu_user_children"",""hostname"":""host_1"",""region"":""eastus""}",0.000000000
-2022-11-22T10:22:05.002Z,-9070404444212865161,"{""measurement"":""used_cpu_user_children"",""hostname"":""host_1"",""region"":""eastus""}",1.000000000
-2022-11-22T10:22:06.003Z,-9070404444212865161,"{""measurement"":""used_cpu_user_children"",""hostname"":""host_1"",""region"":""eastus""}",2.000000000
-`, string(w.Bytes()))
-
-}
-
-func BenchmarkMarshalCSV(b *testing.B) {
-	ts := &prompb.TimeSeries{
-		Labels: []*prompb.Label{
-			{
-				Name:  []byte("__name__"),
-				Value: []byte("__redis__"),
-			},
-			{
-				Name:  []byte("measurement"),
-				Value: []byte("used_cpu_user_children"),
-			},
-			{
-				Name:  []byte("hostname"),
-				Value: []byte("host_1"),
-			},
-			{
-				Name:  []byte("region"),
-				Value: []byte("eastus"),
-			},
-		},
-
-		Samples: []*prompb.Sample{
-			{
-				Timestamp: 1669112524001,
-				Value:     0,
-			},
-			{
-				Timestamp: 1669112525002,
-				Value:     1,
-			},
-			{
-				Timestamp: 1669112526003,
-				Value:     2,
-			},
-		},
-	}
-
-	buf := bytes.NewBuffer(make([]byte, 0, 64*1024))
-	w := NewCSVWriter(buf, []string{"region", "Hostname", "bar"})
-	b.ResetTimer()
-
-	for i := 0; i < b.N; i++ {
-		w.MarshalTS(ts)
-		buf.Reset()
-	}
-}
-
-func TestMarshalCSV_LiftLabel(t *testing.T) {
-	ts := &prompb.TimeSeries{
-		Labels: []*prompb.Label{
-			{
-				Name:  []byte("__name__"),
-				Value: []byte("__redis__"),
-			},
-			{
-				Name:  []byte("hostname"),
-				Value: []byte("host_1"),
-			},
-			{
-				Name:  []byte("measurement"),
-				Value: []byte("used_cpu_user_children"),
-			},
-			{
-				Name:  []byte("region"),
-				Value: []byte("eastus"),
-			},
-		},
-
-		Samples: []*prompb.Sample{
-			{
-				Timestamp: 1669112524001,
-				Value:     0,
-			},
-		},
-	}
-
-	var b bytes.Buffer
-	w := NewCSVWriter(&b, []string{"zip", "zap", "region", "Hostname", "bar"})
-
-	err := w.MarshalTS(ts)
-	require.NoError(t, err)
-	require.Equal(t, `2022-11-22T10:22:04.001Z,1265838189064375029,,host_1,eastus,,,"{""measurement"":""used_cpu_user_children""}",0.000000000
-`, b.String())
-}
-
-func TestNormalize(t *testing.T) {
-	require.Equal(t, "Redis", string(Normalize([]byte("__redis__"))))
-	require.Equal(t, "UsedCpuUserChildren", string(Normalize([]byte("used_cpu_user_children"))))
-	require.Equal(t, "Host1", string(Normalize([]byte("host_1"))))
-	require.Equal(t, "Region", string(Normalize([]byte("region"))))
-	require.Equal(t, "JobEtcdRequestLatency75pctlrate5m", string(Normalize([]byte("Job:etcdRequestLatency:75pctlrate5m"))))
-	require.Equal(t, "TestLimit", string(Normalize([]byte("Test$limit"))))
-	require.Equal(t, "TestRateLimit", string(Normalize([]byte("Test::Rate$limit"))))
-}
 
 func TestTimeConversionForOTLPLogs(t *testing.T) {
 	tests := []struct {
